Removed sensitive environment variables from SAM template.
This is not the right way to handle those.  I'm still not sure what the right way to deploy secrets across many Lambda functions, but this is not the right way.  I was hoping that CloudFormation would copy local environment variables to the Lambda functions on deployment like the SAM CLI apparently does, but that didn't work.