Minor improvement.
The first three tests that fail really should fail, because the authentication logic is wrong.  It should accept either an authentication token in a cookie, or an authorization token in the URL.  The last two HTTP POST tests that are failing probably have problems with both the test and the functionality.